+++ doc/python/conf.py
@@ -7,6 +7,7 @@ sys.path = [os.path.join(os.path.dirname(__file__), '../../build/src/python/')] 
 import corrade
 import corrade.containers
 import corrade.pluginmanager
+import corrade.utility
 
 import magnum
 import magnum.gl
@@ -24,7 +25,7 @@ import magnum.trade
 
 # So the doc see everything
 # TODO: use just +=, m.css should reorder this on its own
-corrade.__all__ = ['containers', 'pluginmanager', 'BUILD_DEPRECATED', 'BUILD_STATIC', 'BUILD_MULTITHREADED', 'TARGET_UNIX', 'TARGET_APPLE', 'TARGET_IOS', 'TARGET_IOS_SIMULATOR', 'TARGET_WINDOWS', 'TARGET_WINDOWS_RT', 'TARGET_EMSCRIPTEN', 'TARGET_ANDROID']
+corrade.__all__ = ['containers', 'pluginmanager', 'utility', 'BUILD_DEPRECATED', 'BUILD_STATIC', 'BUILD_MULTITHREADED', 'TARGET_UNIX', 'TARGET_APPLE', 'TARGET_IOS', 'TARGET_IOS_SIMULATOR', 'TARGET_WINDOWS', 'TARGET_WINDOWS_RT', 'TARGET_EMSCRIPTEN', 'TARGET_ANDROID']
 magnum.__all__ = ['math', 'gl', 'meshtools', 'platform', 'primitives', 'shaders', 'scenegraph', 'text', 'trade', 'BUILD_DEPRECATED', 'BUILD_STATIC', 'TARGET_GL', 'TARGET_GLES', 'TARGET_GLES2', 'TARGET_WEBGL', 'TARGET_EGL', 'TARGET_VK'] + magnum.__all__
 
 # hide values of the preprocessor defines to avoid confusion by assigning a
@@ -176,6 +177,7 @@ INPUT_DOCS = [
     'corrade.rst',
     'corrade.containers.rst',
     'corrade.pluginmanager.rst',
+    'corrade.utility.rst',
 
     'magnum.rst',
     'magnum.gl.rst',